aws-textract/requirements.txt
# requirements.txt

boto3==1.34.61
botocore==1.34.61
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.0
six==1.16.0
textract-trp==0.1.3
urllib3==2.0.7
